+++ Projects/Visualisations/Visualisations/UI_Recorder.h
@@ -1,0 +1,16 @@
+#pragma once
+#include "Recorder.h"
+
+class RecorderUI
+{
+public:
+	RecorderUI(std::shared_ptr<Recorder>& recorder)
+		:
+		m_recorder{ recorder }
+	{
+	}
+
+	void draw() const;
+private:
+	std::shared_ptr<Recorder> m_recorder;
+};
